single file evm config need to clean imports and minimize type aliasing next
